+++ pyk/src/pyk/__main__.py
@@ -7,7 +7,6 @@ from collections.abc import Iterable
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-import tomli
 from graphviz import Digraph
 
 from .cli.pyk import PrintInput, create_argument_parser, generate_options, parse_toml_args
@@ -39,7 +38,6 @@ from .proof.show import APRProofNodePrinter, APRProofShow
 from .utils import check_file_path, ensure_dir_path, exit_with_process_error
 
 if TYPE_CHECKING:
-    from argparse import Namespace
     from typing import Any, Final
 
     from .cli.pyk import (
@@ -383,41 +381,6 @@ def exec_json_to_kore(options: JsonToKoreOptions) -> None:
     kore = Pattern.from_json(text)
     kore.write(sys.stdout)
     sys.stdout.write('\n')
-
-
-def parse_toml_args(args: Namespace) -> dict[str, Any | Iterable]:
-    def get_profile(toml_profile: dict[str, Any], profile_list: list[str]) -> dict[str, Any]:
-        if len(profile_list) == 0 or profile_list[0] not in toml_profile:
-            return {k: v for k, v in toml_profile.items() if type(v) is not dict}
-        elif len(profile_list) == 1:
-            return {k: v for k, v in toml_profile[profile_list[0]].items() if type(v) is not dict}
-        return get_profile(toml_profile[profile_list[0]], profile_list[1:])
-
-    toml_args = {}
-    if args.config_file.is_file():
-        with open(args.config_file, 'rb') as config_file:
-            try:
-                toml_args = tomli.load(config_file)
-            except tomli.TOMLDecodeError:
-                _LOGGER.error(
-                    'Input config file is not in TOML format, ignoring the file and carrying on with the provided command line agruments'
-                )
-
-    toml_args = (
-        get_profile(toml_args[args.command], args.config_profile.split('.')) if args.command in toml_args else {}
-    )
-    toml_args = {get_option_string_destination(args.command, k): v for k, v in toml_args.items()}
-    for k, v in toml_args.items():
-        if k[:3] == 'no-' and (v == 'true' or v == 'false'):
-            del toml_args[k]
-            toml_args[k[3:]] = 'false' if v == 'true' else 'true'
-        if k == 'optimization-level':
-            level = toml_args[k] if toml_args[k] >= 0 else 0
-            level = level if toml_args[k] <= 3 else 3
-            del toml_args[k]
-            toml_args['-O' + str(level)] = 'true'
-
-    return toml_args
 
 
 if __name__ == '__main__':
