+++ pyk/src/pyk/cli/pyk.py
@@ -2,17 +2,24 @@ from __future__ import annotations
 
 import logging
 from argparse import ArgumentParser, FileType
+import logging
+from argparse import ArgumentParser, FileType
 from enum import Enum
 from typing import IO, TYPE_CHECKING, Any
 
 import tomli
 
+import tomli
+
 from ..ktool import TypeInferenceMode
 from ..ktool.kompile import KompileBackend
+from ..ktool import TypeInferenceMode
 from .args import (
+    ConfigArgs,
     ConfigArgs,
     DefinitionOptions,
     DisplayOptions,
+    KCLIArgs,
     KCLIArgs,
     KDefinitionOptions,
     KompileOptions,
@@ -23,14 +30,19 @@ from .args import (
     WarningOptions,
 )
 from .utils import dir_path, file_path
+from .utils import dir_path, file_path
 
 if TYPE_CHECKING:
+    from argparse import Namespace
     from argparse import Namespace
     from collections.abc import Iterable
     from pathlib import Path
     from typing import Final
 
+    from typing import Final
 
+
+_LOGGER: Final = logging.getLogger(__name__)
 _LOGGER: Final = logging.getLogger(__name__)
 
 
@@ -280,7 +292,7 @@ def create_argument_parser() -> ArgumentParser:
     rpc_print_args = pyk_args_command.add_parser(
         'rpc-print',
         help='Pretty-print an RPC request/response',
-        parents=[k_cli_args.logging_args],
+        parents=[k_cli_args.logging_args, config_args.config_args],
     )
     rpc_print_args.add_argument('definition_dir', type=dir_path, help='Path to definition directory.')
     rpc_print_args.add_argument(
@@ -293,7 +305,7 @@ def create_argument_parser() -> ArgumentParser:
     rpc_kast_args = pyk_args_command.add_parser(
         'rpc-kast',
         help='Convert an "execute" JSON RPC response to a new "execute" or "simplify" request, copying parameters from a reference request.',
-        parents=[k_cli_args.logging_args],
+        parents=[k_cli_args.logging_args, config_args.config_args],
     )
     rpc_kast_args.add_argument(
         'reference_request_file',
@@ -310,7 +322,7 @@ def create_argument_parser() -> ArgumentParser:
     prove_legacy_args = pyk_args_command.add_parser(
         'prove-legacy',
         help='Prove an input specification (using kprovex).',
-        parents=[k_cli_args.logging_args],
+        parents=[k_cli_args.logging_args, config_args.config_args],
     )
     prove_legacy_args.add_argument('definition_dir', type=dir_path, help='Path to definition directory.')
     prove_legacy_args.add_argument('main_file', type=str, help='Main file used for kompilation.')
@@ -322,19 +334,14 @@ def create_argument_parser() -> ArgumentParser:
     kompile_args = pyk_args_command.add_parser(
         'kompile',
         help='Kompile the K specification.',
-        parents=[
-            k_cli_args.logging_args,
-            k_cli_args.warning_args,
-            k_cli_args.definition_args,
-            k_cli_args.kompile_args,
-        ],
+        parents=[k_cli_args.logging_args, k_cli_args.definition_args, k_cli_args.kompile_args, config_args.config_args],
     )
     kompile_args.add_argument('main_file', type=str, help='File with the specification module.')
 
     run_args = pyk_args_command.add_parser(
         'run',
         help='Run a given program using the K definition.',
-        parents=[k_cli_args.logging_args],
+        parents=[k_cli_args.logging_args, config_args.config_args],
     )
     run_args.add_argument('pgm_file', type=str, help='File program to run in it.')
     run_args.add_argument('--definition', type=dir_path, dest='definition_dir', help='Path to definition to use.')
@@ -384,7 +391,7 @@ def create_argument_parser() -> ArgumentParser:
     return pyk_args
 
 
-def parse_toml_args(args: Namespace) -> dict[str, Any]:
+def parse_toml_args(args: Namespace) -> dict[str, Any | Iterable]:
     def get_profile(toml_profile: dict[str, Any], profile_list: list[str]) -> dict[str, Any]:
         if len(profile_list) == 0 or profile_list[0] not in toml_profile:
             return {k: v for k, v in toml_profile.items() if type(v) is not dict}
@@ -392,32 +399,28 @@ def parse_toml_args(args: Namespace) -> dict[str, Any]:
             return {k: v for k, v in toml_profile[profile_list[0]].items() if type(v) is not dict}
         return get_profile(toml_profile[profile_list[0]], profile_list[1:])
 
-    toml_args: dict[str, Any] = {}
-    if not hasattr(args, 'config_file') or not args.config_file.is_file():
-        return {}
-
-    with open(args.config_file, 'rb') as config_file:
-        try:
-            toml_args = tomli.load(config_file)
-        except tomli.TOMLDecodeError:
-            _LOGGER.error(
-                'Input config file is not in TOML format, ignoring the file and carrying on with the provided command line agruments'
-            )
+    toml_args = {}
+    if args.config_file.is_file():
+        with open(args.config_file, 'rb') as config_file:
+            try:
+                toml_args = tomli.load(config_file)
+            except tomli.TOMLDecodeError:
+                _LOGGER.error(
+                    'Input config file is not in TOML format, ignoring the file and carrying on with the provided command line agruments'
+                )
 
     toml_args = (
         get_profile(toml_args[args.command], args.config_profile.split('.')) if args.command in toml_args else {}
     )
-
-    toml_adj_args: dict[str, Any] = {}
+    toml_args = {get_option_string_destination(args.command, k): v for k, v in toml_args.items()}
     for k, v in toml_args.items():
-        opt_string = get_option_string_destination(args.command, k)
-        if opt_string[:3] == 'no_':
-            toml_adj_args[opt_string[3:]] = not v
-        elif k == 'optimization-level':
+        if k[:3] == 'no-' and (v == 'true' or v == 'false'):
+            del toml_args[k]
+            toml_args[k[3:]] = 'false' if v == 'true' else 'true'
+        if k == 'optimization-level':
             level = toml_args[k] if toml_args[k] >= 0 else 0
             level = level if toml_args[k] <= 3 else 3
-            toml_adj_args['-o' + str(level)] = True
-        else:
-            toml_adj_args[opt_string] = v
+            del toml_args[k]
+            toml_args['-o' + str(level)] = 'true'
 
-    return toml_adj_args
+    return toml_args
